Phase 2: Decompose SearchEngine into focused modules
- Extract pattern matching into clang_index_mcp/_search/pattern_matcher.py
  (is_pattern, matches, normalize_template_whitespace, detect_pattern_type,
   matches_qualified_pattern).
- Extract prototype building into clang_index_mcp/_search/prototype_builder.py
  (build_function_prototype, build_class_prototype, build_attributes).
- Extract type alias expansion into clang_index_mcp/_search/type_alias_expander.py
  (TypeAliasExpander, expand_type_name).
- Update SearchEngine to delegate to the new modules and remove duplicated code.
- Update QueryEngine and affected tests to import helpers from the new modules.
- Full test suite passes: 1432 passed, 17 skipped, 4 xpassed.
